# openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DDF Select",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://ddf-select.tii.one"
        }
    ],
    "paths": {
        "/api/albums": {
            "get": {
                "tags": [
                    "Albums"
                ],
                "summary": "Get Albums",
                "description": "List all albums.",
                "operationId": "2b1cca5bd2a370a3f7764c5ed249d763",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/albumInclude"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    }
                }
            }
        },
        "/api/albums/{id}": {
            "get": {
                "tags": [
                    "Albums"
                ],
                "summary": "Show Album",
                "description": "Shows a specific album.",
                "operationId": "b8021a5700e53be704e3643f55ee0e1c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The unique identifier of the album",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "$ref": "#/components/parameters/albumInclude"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AlbumResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    }
                }
            }
        },
        "/api/albums/random": {
            "get": {
                "tags": [
                    "Albums"
                ],
                "summary": "Get Random Album",
                "description": "Shows a random album.",
                "operationId": "15a39f315c9136bc2d97bc67f67cb897",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/albumInclude"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AlbumResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    }
                }
            }
        },
        "/api/episodes": {
            "get": {
                "tags": [
                    "Episodes"
                ],
                "summary": "List All Episodes",
                "description": "List all episodes",
                "operationId": "7ec4c2e40b12c6630cbd49f872f4e1b0",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/episodeInclude"
                    },
                    {
                        "$ref": "#/components/parameters/episodeDuration"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    }
                }
            }
        },
        "/api/episodes/{id}": {
            "get": {
                "tags": [
                    "Episodes"
                ],
                "summary": "Get Episode",
                "description": "Get episode",
                "operationId": "63a3e519f54b01e580838484e0fda9a7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The unique identifier of the episode",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "$ref": "#/components/parameters/episodeInclude"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EpisodeResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    }
                }
            }
        },
        "/api/episodes/random": {
            "get": {
                "tags": [
                    "Episodes"
                ],
                "summary": "Get Random Episode,",
                "description": "Get a random episode,",
                "operationId": "9d11e1ba09c9f8b74efe7efb7570b015",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/episodeInclude"
                    },
                    {
                        "$ref": "#/components/parameters/episodeDuration"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EpisodeResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "ArtworkData": {
                "required": [
                    "height",
                    "width",
                    "url"
                ],
                "properties": {
                    "bgColor": {
                        "description": "The average background color of the image.",
                        "type": "string"
                    },
                    "height": {
                        "description": "The maximum height available for the image",
                        "type": "integer"
                    },
                    "width": {
                        "description": "The maximum width available for the image.",
                        "type": "integer"
                    },
                    "textColor1": {
                        "description": "The primary text color used if the background color gets displayed.",
                        "type": "string"
                    },
                    "textColor2": {
                        "description": "The secondary text color used if the background color gets displayed.",
                        "type": "string"
                    },
                    "textColor3": {
                        "description": "The tertiary text color used if the background color gets displayed.",
                        "type": "string"
                    },
                    "textColor4": {
                        "description": "The final post-tertiary text color used if the background color gets displayed.",
                        "type": "string"
                    },
                    "url": {
                        "description": "The URL to request the image asset. {w}x{h}must precede image filename, as placeholders for the width and height values as described above. For example, {w}x{h}bb.jpeg).",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "AlbumResource": {
                "properties": {
                    "id": {
                        "description": "The identifier for the album.",
                        "type": "integer"
                    },
                    "apple_music_id": {
                        "description": "The Apple Music identifier for the album.",
                        "type": "string"
                    },
                    "name": {
                        "description": "The localized name of the album.",
                        "type": "string"
                    },
                    "track_count": {
                        "description": "The number of tracks for the album.",
                        "type": "integer"
                    },
                    "release_date": {
                        "description": "The release date of the album, when known, in YYYY-MM-DD or YYYY format. Prerelease content may have an expected release date in the future.",
                        "type": "string",
                        "nullable": true
                    },
                    "url": {
                        "description": "The URL for sharing the album in Apple Music.",
                        "type": "string"
                    },
                    "artwork": {
                        "$ref": "#/components/schemas/ArtworkData"
                    }
                },
                "type": "object"
            },
            "EpisodeResource": {
                "properties": {
                    "id": {
                        "description": "The identifier for the episode.",
                        "type": "integer"
                    },
                    "number": {
                        "description": "Episode number if the episode has one",
                        "type": "integer",
                        "nullable": true
                    },
                    "name": {
                        "description": "The name of the episode.",
                        "type": "string"
                    },
                    "duration_in_millis": {
                        "description": "",
                        "type": "integer"
                    },
                    "duration_formatted": {
                        "description": "",
                        "type": "string"
                    },
                    "release_date": {
                        "description": "",
                        "type": "string",
                        "format": "YYYY-mm-dd",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "responses": {
            "401": {
                "description": "Unauthorized",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Unauthenticated."
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "404": {
                "description": "Not Found",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Record not found."
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "parameters": {
            "albumInclude": {
                "name": "include",
                "in": "query",
                "description": "Include relationships",
                "style": "form",
                "explode": false,
                "schema": {
                    "type": "array",
                    "items": {
                        "type": "string",
                        "enum": [
                            "episodes",
                            "episodes.tracks"
                        ]
                    }
                }
            },
            "episodeInclude": {
                "name": "include",
                "in": "query",
                "style": "form",
                "explode": false,
                "schema": {
                    "type": "array",
                    "items": {
                        "type": "string",
                        "enum": [
                            "album",
                            "tracks"
                        ]
                    }
                }
            },
            "episodeDuration": {
                "name": "filter[duration]",
                "in": "query",
                "schema": {
                    "type": "string",
                    "enum": [
                        "short",
                        "normal",
                        "long"
                    ]
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "apiKey",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "tags": [
        {
            "name": "Albums",
            "description": "Endpoints related to albums"
        },
        {
            "name": "Episodes",
            "description": "Endpoints related to episodes"
        }
    ]
}